@@ -1,10 +1,9 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 
 import PackageDescription
 
 let settings: [SwiftSetting] = [.unsafeFlags(["-warnings-as-errors"]),
-                                .enableExperimentalFeature("ExistentialAny"),
-                                .enableExperimentalFeature("StrictConcurrency=targeted")]
+                                .enableExperimentalFeature("ExistentialAny")]
 
 let package = Package(
     name: "MEGATest",
@@ -27,5 +26,6 @@ let package = Package(
             ],
             swiftSettings: settings
         )
-    ]
+    ],
+    swiftLanguageModes: [.v6]
 )
